updating telegram chat and channel API to reflect 133 API layer

+++ tl/src/main/java/com/github/badoualy/telegram/tl/api/request/TLRequestMessagesGetFullChat.java
@@ -78,7 +78,7 @@ public class TLRequestMessagesGetFullChat extends TLMethod<TLChatFull> {
         return chatId;
     }
 
-    public void setChatId(int chatId) {
+    public void setChatId(long chatId) {
         this.chatId = chatId;
     }
 }
